Structure changes.
