+++ CHANGELOG.md
@@ -1,0 +1,9 @@
+# Changelog
+
+## 1.0.0
+- Refactored entire code base.
+- Validator
+  - Removed `checkUser`.
+  - Removed `checkPassword`.
+  - Removed `checkVersion`.
+  - Removed `checkId`.
