+++ lib/Validator.js
@@ -18,7 +18,7 @@
 
 const check = require('check-types');
 
-const messages = require('./messageContainer');
+const messages = require('./provider/messages');
 const NoArgumentError = require('./error/NoArgumentError');
 const InvalidArgumentError = require('./error/InvalidArgumentError');
 
@@ -39,8 +39,7 @@ class Validator {
   _checkTypeArgument(type) {
     // Check type of `type` argument, if provided.
     if (check.assigned(type) && check.not.string(type)) {
-      const message = `Type argument ${messages.IS_NOT_STRING}`;
-      throw new InvalidArgumentError(message);
+      throw new InvalidArgumentError(`Type argument ${messages.IS_NOT_STRING}`);
     }
   }
 
@@ -48,173 +47,61 @@ class Validator {
    * Checks an object argument for existence and type.
    *
    * @param  {object} objectParam   Object parameter provided.
-   * @param  {string} [type]        Object entitiy name.
-   * Used if the entity's name should appear in error.
+   * @param  {string} [type]        Object entity name.
    *
-   * @throws NoArgumentError      On missing object.
-   * @throws InvalidArgumentError On wrong type.
+   * @throws NoArgumentError        On missing object.
+   * @throws InvalidArgumentError   On wrong type.
    *
    * @example
-   * 	// Throws 'InvalidArgumentError: Server credentials is not a string'.
+   * 	// Throws 'InvalidArgumentError: Server credentials is not an object'.
    * 	checkObject('asdsad', 'Server credentials');
+   *
+   * @static
    */
-  checkObject(objectParam, type) {
+  static checkObject(object, type) {
     // Check and assign object entity name for error message.
     this._checkTypeArgument(type);
-    const objectEntity = check.assigned(type) ? type : messages.PARAM;
+    const objectEntity = check.assigned(type) ? type : 'Object';
 
     // Check existence.
-    if (check.not.assigned(objectParam)) {
-      const message = `${objectEntity} ${messages.IS_MISSING}`;
-      throw new NoArgumentError(message);
+    if (check.not.assigned(object)) {
+      throw new NoArgumentError(`${objectEntity} ${messages.IS_MISSING}`);
     }
 
     // Check type.
-    if (check.not.object(objectParam)) {
-      const message = `${objectEntity} ${messages.IS_NOT_OBJECT}`;
-      throw new InvalidArgumentError(message);
+    if (check.not.object(object)) {
+      throw new InvalidArgumentError(`${objectEntity} ${messages.IS_NOT_OBJECT}`);
     }
   }
 
   /**
    * Checks an URL argument for existence, type and format validity.
    *
-   * @param  {string} url URL provided.
+   * @param {string} url URL provided.
    *
    * @throws NoArgumentError      On missing URL.
    * @throws InvalidArgumentError On wrong type and invalid URL format.
+   *
+   * @static
    */
-  checkUrl(url) {
+  static checkUrl(url) {
     // Regular expression to test the URL format against.
     const formatRegex = /https?:\/\/(?:www\.|(?!www))[^\s\.]+\.[^\s]{2,}|www\.[^\s]+\.[^\s]{2,}/;
 
+    this.checkString();
     // Check existence.
     if (check.not.assigned(url)) {
-      const message = `${messages.URL} ${messages.IS_MISSING}`;
-      throw new NoArgumentError(message);
+      throw new NoArgumentError(`URL ${messages.IS_MISSING}`);
     }
 
     // Check type.
     if (check.not.string(url)) {
-      const message = `${messages.URL} ${messages.IS_NOT_STRING}`;
-      throw new InvalidArgumentError(message);
+      throw new InvalidArgumentError(`URL ${messages.IS_NOT_STRING}`);
     }
 
     // Check format.
     if (check.not.match(url, formatRegex)) {
-      const message = `${messages.URL} ${messages.IS_INVALID}`;
-      throw new InvalidArgumentError(message);
-    }
-  }
-
-  /**
-   * Checks an user argument for existence and type.
-   *
-   * @param  {string} user User provided.
-   *
-   * @throws NoArgumentError      On missing User.
-   * @throws InvalidArgumentError On wrong type.
-   */
-  checkUser(user) {
-    // Check existence.
-    if (check.not.assigned(user)) {
-      const message = `${messages.USER} ${messages.IS_MISSING}`;
-      throw new NoArgumentError(message);
-    }
-
-    // Check type.
-    if (check.not.string(user)) {
-      const message = `${messages.USER} ${messages.IS_NOT_STRING}`;
-      throw new InvalidArgumentError(message);
-    }
-  }
-
-  /**
-   * Checks a password argument for existence and type.
-   *
-   * @param  {string} password Password provided.
-   *
-   * @throws NoArgumentError      On missing Password.
-   * @throws InvalidArgumentError On wrong type.
-   */
-  checkPassword(password) {
-    // Check existence.
-    if (check.not.assigned(password)) {
-      const message = `${messages.PASSWORD} ${messages.IS_MISSING}`;
-      throw new NoArgumentError(message);
-    }
-
-    // Check type.
-    if (check.not.string(password)) {
-      const message = `${messages.PASSWORD} ${messages.IS_NOT_STRING}`;
-      throw new InvalidArgumentError(message);
-    }
-  }
-
-  /**
-   * Checks an ID argument for existence and type
-   * and validates if it is an integer.
-   *
-   * @param  {integer} id     ID provided.
-   * @param  {string} [type]  ID entitiy name. Used if the entity's name should appear in error.
-   *
-   * @throws NoArgumentError      On missing Password.
-   * @throws InvalidArgumentError On wrong type.
-   *
-   * @example
-   * 	// Throws 'InvalidArgumentError: Country ID is not a string'.
-   * 	checkId('asdsad', 'Country ID');
-   */
-  checkId(id, type) {
-    // Check and assign ID entity name for error message.
-    this._checkTypeArgument(type);
-    const idEntity = check.assigned(type) ? type : messages.ID;
-
-    // Check existence.
-    if (check.not.assigned(id)) {
-      const message = `${idEntity} ${messages.IS_MISSING}`;
-      throw new NoArgumentError(message);
-    }
-
-    // Check type.
-    if (check.not.number(id)) {
-      const message = `${idEntity} ${messages.IS_NOT_NUMBER}`;
-      throw new InvalidArgumentError(message);
-    }
-
-    // Check if it is an integer.
-    if (check.not.integer(id)) {
-      const message = `${idEntity} ${messages.IS_NOT_INTEGER}`;
-      throw new InvalidArgumentError(message);
-    }
-  }
-
-  /**
-   * Checks an API version argument for existence, type and format validity.
-   *
-   * @param  {string} version API version provided.
-   *
-   * @throws NoArgumentError      On missing Password.
-   * @throws InvalidArgumentError On wrong type.
-   */
-  checkVersion(version) {
-    // Check existence.
-    if (check.not.assigned(version)) {
-      const message = `${messages.VERSION} ${messages.IS_MISSING}`;
-      throw new NoArgumentError(message);
-    }
-
-    // Check type.
-    if (check.not.string(version)) {
-      const message = `${messages.VERSION} ${messages.IS_NOT_STRING}`;
-      throw new InvalidArgumentError(message);
-    }
-
-    // Check format.
-    // String should always start with 'v'.
-    if (version.charAt(0) !== 'v') {
-      const message = `${messages.VERSION} ${messages.IS_INVALID}`;
-      throw new InvalidArgumentError(message);
+      throw new InvalidArgumentError(`URL ${messages.IS_INVALID}`);
     }
   }
