--- lib/messageContainer.js
@@ -1,47 +1,0 @@
-/**
- * @name Message container.
- *
- * @description Contains message strings.
- *
- * @type {Object}
- */
-module.exports = {
-  // Suffixes.
-  IS_NOT_STRING: 'is not a string',
-  IS_NOT_OBJECT: 'is not an object',
-  IS_NOT_INTEGER: 'is not an integer',
-  IS_NOT_NUMBER: 'is not a number',
-  IS_INVALID: 'is invalid',
-  IS_MISSING: 'is missing',
-
-  // Argument entities.
-  URL: 'URL',
-  USER: 'User',
-  PASSWORD: 'Password',
-  CREDENTIALS: 'Credentials object',
-  VERSION: 'Version',
-  HEADERS: 'Headers',
-  PARAM: 'Parameter',
-  ID: 'ID',
-  POST_DATA: 'Post data',
-  PUT_DATA: 'Put request data',
-  SEARCH_TERM: 'Search term',
-  FILE: 'File',
-  REQUEST_METHOD: 'HTTP request method',
-  FORM_DATA: 'Form data',
-
-  // Primitives.
-  NUMBER: 'Number',
-  STRING: 'String',
-
-  // API entities.
-  COUNTRY_ID: 'Country ID',
-  ZONE_ID: 'Zone ID',
-  ADDRESS_ID: 'Address ID',
-  ADDRESS_DATA: 'Address data',
-  SORTING_OBJECT: 'Sorting object',
-  CUSTOMER_ID: 'Customer ID',
-  CUSTOMER_DATA: 'Customer data',
-  EMAIL_ID: 'E-Mail ID',
-  EMAIL_DATA: 'E-Mail data',
-};
