+++ lib/provider/messages.js
@@ -1,0 +1,18 @@
+/**
+ * @name Message provider.
+ *
+ * @description Contains message strings.
+ *
+ * @type {Object}
+ */
+module.exports = {
+  // Type checking.
+  IS_NOT_STRING: 'is not a string',
+  IS_NOT_OBJECT: 'is not an object',
+  IS_NOT_INTEGER: 'is not an integer',
+  IS_NOT_NUMBER: 'is not a number',
+
+  // State checking.
+  IS_INVALID: 'is invalid',
+  IS_MISSING: 'is missing',
+};
