+++ test/error/ExtendableErrorTest.js
